map: Add Test Ground 7
This map was created by me in AlternativaEditor in 3 hours.

I hate this editor, it is basically impossible to do anything in
it without already having concept art of what it should look like.
I'd love to see a Blender addon that could automatically tile
and randomize ground and wall props, so I could spend time on level design
and not on moving tiles just to see that they're at the wrong height.